fix: include metadata labels in selectors
